Reorg tests into /tests directory, and reframe as integration tests
